created utils folder and removed inplace operations
