+++ crates/prover/src/components/opcodes/jnz_fp_imm.rs
@@ -56,7 +56,8 @@ use stwo_prover::core::poly::circle::CircleEvaluation;
 
 use crate::adapter::ExecutionBundle;
 use crate::components::Relations;
-use crate::utils::{Enabler, PackedExecutionBundle};
+use crate::utils::enabler::Enabler;
+use crate::utils::execution_bundle::PackedExecutionBundle;
 
 const N_TRACE_COLUMNS: usize = 9;
 const N_MEMORY_LOOKUPS: usize = 4;
