+++ crates/prover/src/components/opcodes/mod.rs
@@ -192,9 +192,7 @@ define_opcodes! {
     (Opcode::JnzFpImm, jnz_fp_imm_taken),
     (Opcode::Ret, ret),
     (Opcode::StoreAddFpFp, store_add_fp_fp),
-    (Opcode::StoreAddFpFp, store_add_fp_fp_inplace),
     (Opcode::StoreAddFpImm, store_add_fp_imm),
-    (Opcode::StoreAddFpImm, store_add_fp_imm_inplace),
     (Opcode::StoreDerefFp, store_deref_fp),
     (Opcode::StoreDivFpFp, store_div_fp_fp),
     (Opcode::StoreDivFpImm, store_div_fp_imm),
