+++ crates/prover/src/components/opcodes/store_div_fp_fp.rs
@@ -66,7 +66,8 @@ use stwo_prover::core::poly::circle::CircleEvaluation;
 
 use crate::adapter::ExecutionBundle;
 use crate::components::Relations;
-use crate::utils::{Enabler, PackedExecutionBundle};
+use crate::utils::enabler::Enabler;
+use crate::utils::execution_bundle::PackedExecutionBundle;
 
 const N_TRACE_COLUMNS: usize = 15;
 const N_MEMORY_LOOKUPS: usize = 8;
