+++ crates/prover/src/debug_tools/relation_tracker.rs
@@ -107,9 +107,7 @@ fn relation_entries(
         add_to_relation_entries(&opcodes.jnz_fp_imm_taken, trace),
         add_to_relation_entries(&opcodes.ret, trace),
         add_to_relation_entries(&opcodes.store_add_fp_fp, trace),
-        add_to_relation_entries(&opcodes.store_add_fp_fp_inplace, trace),
         add_to_relation_entries(&opcodes.store_add_fp_imm, trace),
-        add_to_relation_entries(&opcodes.store_add_fp_imm_inplace, trace),
         add_to_relation_entries(&opcodes.store_deref_fp, trace),
         add_to_relation_entries(&opcodes.store_div_fp_fp, trace),
         add_to_relation_entries(&opcodes.store_div_fp_imm, trace),
