+++ crates/prover/src/utils/enabler.rs
@@ -1,0 +1,33 @@
+use num_traits::{One, Zero};
+use stwo_prover::core::backend::simd::m31::{N_LANES, PackedM31};
+use stwo_prover::core::fields::m31::M31;
+
+/// The enabler column is a column of length `padding_offset.next_power_of_two()` where
+/// 1. The first `padding_offset` elements are set to 1;
+/// 2. The rest are set to 0.
+#[derive(Debug, Clone)]
+pub struct Enabler {
+    pub padding_offset: usize,
+}
+impl Enabler {
+    pub const fn new(padding_offset: usize) -> Self {
+        Self { padding_offset }
+    }
+
+    pub fn packed_at(&self, vec_row: usize) -> PackedM31 {
+        let row_offset = vec_row * N_LANES;
+        if row_offset >= self.padding_offset {
+            return PackedM31::zero();
+        }
+        if row_offset + N_LANES <= self.padding_offset {
+            return PackedM31::one();
+        }
+
+        // The row is partially enabled.
+        let mut res = [M31::zero(); N_LANES];
+        for v in res.iter_mut().take(self.padding_offset - row_offset) {
+            *v = M31::one();
+        }
+        PackedM31::from_array(res)
+    }
+}
